+++ activiti-cloud-audit-service/activiti-cloud-services-audit-consumer/src/main/java/org/activiti/cloud/services/audit/jpa/streams/config/AuditJPAStreamsAutoConfiguration.java
@@ -30,22 +30,24 @@ import org.activiti.cloud.services.audit.service.TeamsChatService;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
 import org.springframework.messaging.Message;
 
 @AutoConfiguration
 public class AuditJPAStreamsAutoConfiguration {
 
-    @Value("${microsoft.app.id}")
+    @Value("${microsoft.app.id:}")
     private String appId;
 
-    @Value("${microsoft.app.password}")
+    @Value("${microsoft.app.password:}")
     private String appPassword;
 
-    @Value("${teams.connector.tenant}")
+    @Value("${teams.connector.tenant:}")
     private String appTenant;
 
     @Bean
+    @ConditionalOnProperty(name = "teams.connector.chat.enabled", havingValue = "true")
     public TeamsChatService teamsChatService() {
         return new TeamsChatService(appId, appPassword, appTenant);
     }
