+++ activiti-cloud-audit-service/activiti-cloud-services-audit-consumer/src/main/resources/application.properties
@@ -1,0 +1,4 @@
+microsoft.app.id=${MS_APP_ID}
+microsoft.app.password=${MS_APP_PASSWORD}
+teams.connector.tenant=${TEAMS_TENANT}
+teams.connector.chat.enabled=${TEAMS_CHAT_ENABLED:false}
