+++ activiti-cloud-audit-service/activiti-cloud-services-audit-consumer/src/test/java/org/activiti/cloud/services/audit/jpa/streams/AuditConsumerChannelHandlerImplTest.java
@@ -52,16 +52,11 @@ public class AuditConsumerChannelHandlerImplTest {
     @Captor
     private ArgumentCaptor<Iterable<AuditEventEntity>> argumentCaptor;
 
-    //    @Test
+    @Test
     public void receiveEventShouldStoreEntity() {
         //given
         CloudRuntimeEvent cloudRuntimeEvent = Mockito.mock(CloudRuntimeEventImpl.class);
         Mockito.when(cloudRuntimeEvent.getEventType()).thenReturn(ProcessRuntimeEvent.ProcessEvents.PROCESS_CREATED);
-        Mockito.when(cloudRuntimeEvent.getAppName()).thenReturn("TestApp");
-        Mockito.when(cloudRuntimeEvent.getActor()).thenReturn(getJsonDetails());
-        Mockito.when(cloudRuntimeEvent.getTimestamp()).thenReturn(new Date().getTime());
-        Mockito.when(cloudRuntimeEvent.getProcessInstanceId()).thenReturn("ad4f8843-bb25-11f0-bebc-1ac6710786eb");
-        Mockito.when(cloudRuntimeEvent.getId()).thenReturn("ad546a47-bb25-11f0-bebc-1ac6710786eb");
 
         EventToEntityConverter converter = Mockito.mock(EventToEntityConverter.class);
         Mockito
@@ -87,12 +82,11 @@ public class AuditConsumerChannelHandlerImplTest {
         Assertions.assertThat(argumentCaptor.getValue()).containsOnly(entity);
     }
 
-    //    @Test
+    @Test
     public void messageIdShouldBeSet() {
         //given
         CloudRuntimeEvent cloudRuntimeEvent = Mockito.mock(CloudRuntimeEventImpl.class);
         Mockito.when(cloudRuntimeEvent.getEventType()).thenReturn(ProcessRuntimeEvent.ProcessEvents.PROCESS_CREATED);
-        Mockito.when(cloudRuntimeEvent.getAppName()).thenReturn("TestApp");
         EventToEntityConverter converter = Mockito.mock(EventToEntityConverter.class);
         Mockito
             .when(converters.getConverterByEventTypeName(ProcessRuntimeEvent.ProcessEvents.PROCESS_CREATED.name()))
@@ -110,190 +104,5 @@ public class AuditConsumerChannelHandlerImplTest {
 
         //then
         Mockito.verify((CloudRuntimeEventImpl) cloudRuntimeEvent).setMessageId(headers.get("id").toString());
-    }
-
-    private String getJsonDetails() {
-        return """
-            {
-              "entity": {
-                "id": "ad546a47-bb25-11f0-bebc-1ac6710786eb",
-                "inBoundVariables": {
-                  "restUrl": {
-                    "type": "string",
-                    "value": "https://aaaaa.com"
-                  }
-                },
-                "outBoundVariables": {},
-                "processInstanceId": "ad4f8843-bb25-11f0-bebc-1ac6710786eb",
-                "rootProcessInstanceId": "ad4f8843-bb25-11f0-bebc-1ac6710786eb",
-                "processDefinitionId": "Process_1762442098730:1:a51db671-bb24-11f0-bebc-1ac6710786eb",
-                "executionId": "ad502486-bb25-11f0-bebc-1ac6710786eb",
-                "processDefinitionKey": "Process_1762442098730",
-                "processDefinitionVersion": 1,
-                "clientId": "restConnector_1o6p9s0",
-                "clientType": "ServiceTask",
-                "appVersion": "1",
-                "connectorType": "rest-connector-e0r2c.POST"
-              },
-              "errorMessage": "No subject alternative DNS name matching aaaaa.com found.",
-              "stackTraceElements": [
-                {
-                  "methodName": "matchDNS",
-                  "fileName": "HostnameChecker.java",
-                  "lineNumber": 207,
-                  "nativeMethod": false,
-                  "className": "sun.security.util.HostnameChecker"
-                },
-                {
-                  "methodName": "match",
-                  "fileName": "HostnameChecker.java",
-                  "lineNumber": 103,
-                  "nativeMethod": false,
-                  "className": "sun.security.util.HostnameChecker"
-                },
-                {
-                  "methodName": "checkIdentity",
-                  "fileName": "X509TrustManagerImpl.java",
-                  "lineNumber": 466,
-                  "nativeMethod": false,
-                  "className": "sun.security.ssl.X509TrustManagerImpl"
-                },
-                {
-                  "methodName": "checkIdentity",
-                  "fileName": "X509TrustManagerImpl.java",
-                  "lineNumber": 417,
-                  "nativeMethod": false,
-                  "className": "sun.security.ssl.X509TrustManagerImpl"
-                },
-                {
-                  "methodName": "checkTrusted",
-                  "fileName": "X509TrustManagerImpl.java",
-                  "lineNumber": 237,
-                  "nativeMethod": false,
-                  "className": "sun.security.ssl.X509TrustManagerImpl"
-                },
-                {
-                  "methodName": "checkServerTrusted",
-                  "fileName": "X509TrustManagerImpl.java",
-                  "lineNumber": 132,
-                  "nativeMethod": false,
-                  "className": "sun.security.ssl.X509TrustManagerImpl"
-                },
-                {
-                  "methodName": "checkServerCerts",
-                  "fileName": "CertificateMessage.java",
-                  "lineNumber": 631,
-                  "nativeMethod": false,
-                  "className": "sun.security.ssl.CertificateMessage$T12CertificateConsumer"
-                },
-                {
-                  "methodName": "onCertificate",
-                  "fileName": "CertificateMessage.java",
-                  "lineNumber": 467,
-                  "nativeMethod": false,
-                  "className": "sun.security.ssl.CertificateMessage$T12CertificateConsumer"
-                },
-                {
-                  "methodName": "consume",
-                  "fileName": "CertificateMessage.java",
-                  "lineNumber": 363,
-                  "nativeMethod": false,
-                  "className": "sun.security.ssl.CertificateMessage$T12CertificateConsumer"
-                },
-                {
-                  "methodName": "consume",
-                  "fileName": "SSLHandshake.java",
-                  "lineNumber": 393,
-                  "nativeMethod": false,
-                  "className": "sun.security.ssl.SSLHandshake"
-                },
-                {
-                  "methodName": "dispatch",
-                  "fileName": "HandshakeContext.java",
-                  "lineNumber": 477,
-                  "nativeMethod": false,
-                  "className": "sun.security.ssl.HandshakeContext"
-                },
-                {
-                  "methodName": "dispatch",
-                  "fileName": "HandshakeContext.java",
-                  "lineNumber": 448,
-                  "nativeMethod": false,
-                  "className": "sun.security.ssl.HandshakeContext"
-                },
-                {
-                  "methodName": "dispatch",
-                  "fileName": "TransportContext.java",
-                  "lineNumber": 206,
-                  "nativeMethod": false,
-                  "className": "sun.security.ssl.TransportContext"
-                },
-                {
-                  "methodName": "decode",
-                  "fileName": "SSLTransport.java",
-                  "lineNumber": 172,
-                  "nativeMethod": false,
-                  "className": "sun.security.ssl.SSLTransport"
-                },
-                {
-                  "methodName": "decode",
-                  "fileName": "SSLSocketImpl.java",
-                  "lineNumber": 1506,
-                  "nativeMethod": false,
-                  "className": "sun.security.ssl.SSLSocketImpl"
-                },
-                {
-                  "methodName": "readHandshakeRecord",
-                  "fileName": "SSLSocketImpl.java",
-                  "lineNumber": 1421,
-                  "nativeMethod": false,
-                  "className": "sun.security.ssl.SSLSocketImpl"
-                },
-                {
-                  "methodName": "startHandshake",
-                  "fileName": "SSLSocketImpl.java",
-                  "lineNumber": 455,
-                  "nativeMethod": false,
-                  "className": "sun.security.ssl.SSLSocketImpl"
-                },
-                {
-                  "methodName": "startHandshake",
-                  "fileName": "SSLSocketImpl.java",
-                  "lineNumber": 426,
-                  "nativeMethod": false,
-                  "className": "sun.security.ssl.SSLSocketImpl"
-                },
-                {
-                  "methodName": "executeHandshake",
-                  "fileName": "AbstractClientTlsStrategy.java",
-                  "lineNumber": 253,
-                  "nativeMethod": false,
-                  "className": "org.apache.hc.client5.http.ssl.AbstractClientTlsStrategy"
-                },
-                {
-                  "methodName": "upgrade",
-                  "fileName": "AbstractClientTlsStrategy.java",
-                  "lineNumber": 211,
-                  "nativeMethod": false,
-                  "className": "org.apache.hc.client5.http.ssl.AbstractClientTlsStrategy"
-                },
-                {
-                  "methodName": "upgrade",
-                  "fileName": "DefaultClientTlsStrategy.java",
-                  "lineNumber": 48,
-                  "nativeMethod": false,
-                  "className": "org.apache.hc.client5.http.ssl.DefaultClientTlsStrategy"
-                },
-                {
-                  "methodName": "connect",
-                  "fileName": "DefaultHttpClientConnectionOperator.java",
-                  "lineNumber": 219,
-                  "nativeMethod": false,
-                  "className": "org.apache.hc.client5.http.impl.io.DefaultHttpClientConnectionOperator"
-                }
-              ],
-              "errorClassName": "org.activiti.cloud.api.process.model.CloudBpmnError"
-            }
-            """;
     }
 }
