+++ activiti-cloud-audit-service/activiti-cloud-starter-audit-consumer/src/test/java/org/activiti/cloud/starter/audit/consumer/test/ActivitiAuditConsumerAutoConfigurationTest.java
@@ -24,7 +24,6 @@ import org.springframework.cloud.stream.binder.test.EnableTestBinder;
 @SpringBootTest
 @SpringBootApplication
 @EnableTestBinder
-@Disabled
 public class ActivitiAuditConsumerAutoConfigurationTest {
 
     @Test
